+++ packages/Utd/Gifts/database/migrations/2024_01_16_135304_create_user_lucky_gifts_table.php
@@ -6,26 +6,24 @@ use Illuminate\Support\Facades\Schema;
 
 return new class extends Migration
 {
-    /**
-     * Run the migrations.
-     */
     public function up(): void
     {
-        Schema::create('user_lucky_gifts', function (Blueprint $table) {
-            $table->id();
-            $table->foreignId('user_id')->nullable()->constrained('users')->onDelete('cascade');
-            $table->foreignId('gift_id')->nullable();
-            $table->boolean('type')->default(0);
-            $table->double('value')->default(0);
-            $table->integer('number')->default(0);
-            $table->double('gift_price')->default(0);
-            $table->timestamps();
-        });
+        if (! Schema::hasTable('user_lucky_gifts')) {
+            Schema::create('user_lucky_gifts', function (Blueprint $table) {
+                $table->id();
+                $table->foreignId('user_id')->nullable()->constrained('users')->onDelete('cascade');
+                $table->foreignId('gift_id')->nullable();
+                $table->boolean('type')->default(0);
+                $table->double('value')->default(0);
+                $table->bigInteger('number')->default(0);
+                $table->double('gift_price')->default(0);
+                $table->bigInteger('total_win')->default(0);
+                $table->bigInteger('total_num_win')->default(0);
+                $table->timestamps();
+            });
+        }
     }
 
-    /**
-     * Reverse the migrations.
-     */
     public function down(): void
     {
         Schema::dropIfExists('user_lucky_gifts');
